refactor(unified-sdk-jni): hoist net_from_ord into support.rs
Follow-up to #4187 review: tx_decode.rs added the third private copy of
the FFINetwork ordinal mapping (after transactions.rs and identity.rs).
Hoist a single shared copy into support.rs and update all callers —
including a fourth copy in persistence.rs with identical semantics.
rs_sdk_ffi::FFINetwork and platform_wallet_ffi::FFINetwork are both
re-exports of dash_network::ffi::FFINetwork, so the one helper serves
every module. The exhaustive ordinal test moves to support.rs with it.
